backend/passport/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
pymysql==1.1.0
redis==5.0.1
authlib==1.3.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2
python-multipart==0.0.6
email-validator==2.1.0
httpx==0.25.2
captcha==0.5.0
Pillow>=10.0.0
